Beam Search Phone to Grapheme Converter
This version converts to ARPAbet phonemes to words. It utilizes parallelism to perform exhaustic search of phonemes from the cmudict library. The cmudict library has been hashed for O(1) search complexity time. Beam search size is equal to the number of logical processors. The beam search uses the gpt2 model from transformers library to score sentences to ascertain the most accurate conversion. The current approach is not optimal